cdk-infrastructure/lambda/requirements.txt
# Dixon Smart Repair - Strands Agent Dependencies with Direct Tavily API + VIN Processing
# Based on official Strands documentation

# Core Strands Agents SDK
strands-agents>=0.1.0
# Strands Tools Package (contains image_reader and other tools)
strands-agents-tools>=0.1.0

# AWS SDK for Python (required for Bedrock + Textract)
boto3>=1.34.0
botocore>=1.34.0

# HTTP requests for direct Tavily API calls + NHTSA API
requests>=2.31.0

# Image processing for VIN extraction validation
Pillow>=9.0.0

# JSON handling and utilities
python-dateutil>=2.8.2
